Seperate scrapers, pull in podium and championship titles
